refactor: consolidate identity profile into settings, standardize close buttons, and improve graph visualization logic

+++ monorepo/apps/gnosi/backend/api/scheduler_routes.py
@@ -10,7 +10,7 @@ router = APIRouter(prefix="/api/schedulers", tags=["schedulers"])
 
 
 class TaskUpdate(BaseModel):
-    interval_minutes: int
+    interval_minutes: float
     enabled: bool
